Feat: 해당 일 레코드 운동 모두 가져오기 기능 구현 (Repository)

+++ app/src/main/java/com/jaemin/fitzam/data/source/local/entity/WorkoutSetEntity.kt
@@ -21,7 +21,7 @@ import androidx.room.PrimaryKey
 data class WorkoutSetEntity(
     val entryId: Long,
     val setIndex: Int,
-    val weightKg: Float,
+    val weightKg: Double,
     val reps: Int,
     val createdAt: Long,
 )
